requirements.txt
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
astroid==2.0.4
attrs==18.1.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==2.1.3
boto==2.49.0
boto3==1.7.80
botocore==1.10.80
bz2file==0.98
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
click==6.7
click-plugins==1.0.3
cligj==0.4.0
colorama==0.3.7
constantly==15.1.0
cryptography==2.3
cryptography-vectors==2.3
cycler==0.10.0
decorator==4.3.0
docutils==0.14
entrypoints==0.2.3
Fiona==1.7.13
GDAL==2.3.0
gensim==3.5.0
geopandas==0.4.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
jupyterlab==0.33.8
jupyterlab-launcher==0.11.2
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
munch==2.3.2
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.3
notebook==5.6.0
numpy==1.15.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prometheus-client==0.3.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
Pygments==2.2.0
pylint==2.1.1
pymongo==3.7.1
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyproj==1.9.5.1
python-dateutil==2.7.3
pytz==2018.5
pyzmq==17.1.2
redis==2.10.6
requests==2.19.1
s3transfer==0.1.12
scipy==1.1.0
seaborn==0.9.0
selenium==3.14.0
Send2Trash==1.5.0
service-identity==17.0.0
Shapely==1.6.4.post2
simplegeneric==0.8.1
six==1.11.0
smart-open==1.6.0
terminado==0.8.1
testpath==0.3.1
textblob==0.15.1
tornado==5.1
tqdm==4.25.0
traitlets==4.3.2
Twisted==17.5.0
typed-ast==1.1.0
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.10.11
zope.interface==4.5.0
